created util.py, moved Button to there
