+++ User_Interface/MenuView.py
@@ -3,7 +3,7 @@
 # It allows players to navigate and make selections within the game. 
 
 import pygame, os
-from User_Interface.Button import *
+from User_Interface.util import *
 pygame.init()
 
 # Display Choice 
@@ -23,14 +23,12 @@ def main_menu(SCREEN: pygame.Surface) -> int:
     # Sets the Background Image and Sound, Sound loops until Screen transision.
     SCREEN.fill("white")
     bg_image = pygame.transform.smoothscale(pygame.image.load(os.path.join("assets", "images", "title_bg.png")), (SCREEN.get_width(), SCREEN.get_height()))
-    #bg_image.set_alpha(128)
     
     pygame.mixer.init()
     pygame.mixer.music.load(os.path.join("assets","sounds","title_bg_music.mp3"))
     pygame.mixer.music.set_volume(0.45)
     pygame.mixer.music.play(loops= -1)
 
-    title_font = pygame.font.SysFont("consolas", 75)
     button_font =  pygame.font.SysFont("consolas", 50)
 
     # Screen Game Loop, Renders Buttons and Checks if the user clicks them
@@ -41,33 +39,25 @@ def main_menu(SCREEN: pygame.Surface) -> int:
         SCREEN.blit(bg_image,(0,0))
         mouse_pos = pygame.mouse.get_pos()
 
-        menu_text = title_font.render("Cloneopoly", True, "black")
-        menu_text_rect = menu_text.get_rect(center = (SCREEN.get_width()/2, SCREEN.get_height()/5))
-
-        menu_text = title_font.render("", True, "black")
-        menu_text_rect = menu_text.get_rect(center = (SCREEN.get_size()[0]/2, SCREEN.get_size()[1]/5))
-
-      
-        play_button = Button(pos=(SCREEN.get_width()/2, SCREEN.get_height()/3), text_input= "Play", font= button_font , base_color="#39FF14", hover_color= "#0a18f3")
-        options_button = Button(pos=(SCREEN.get_width()/2, SCREEN.get_height()/3 + 100), text_input= "Options", font= button_font, base_color= "#39FF14", hover_color= "#0a18f3")
-        quit_button = Button(pos=(SCREEN.get_width()/2, SCREEN.get_height()/3 + 200), text_input= "Quit", font= button_font, base_color= "#39FF14", hover_color= "#0a18f3")
+        play_button = Button(pos=(SCREEN.get_width()/2, SCREEN.get_height()/3), text_input= "Play", font= button_font , base_color="#000000", hover_color= "#0a18f3")
+        options_button = Button(pos=(SCREEN.get_width()/2, SCREEN.get_height()/3 + 100), text_input= "Options", font= button_font, base_color= "#000000", hover_color= "#0a18f3")
+        quit_button = Button(pos=(SCREEN.get_width()/2, SCREEN.get_height()/3 + 200), text_input= "Quit", font= button_font, base_color= "#000000", hover_color= "#0a18f3")
         
-        SCREEN.blit(menu_text,menu_text_rect)
       
         for button in [play_button, options_button, quit_button]:
-            button.changeColor(mouse_pos)
+            button.change_color(mouse_pos)
             button.update(SCREEN)
         
         for event in pygame.event.get():
             if event.type == pygame.QUIT:
                 run = False
             if event.type == pygame.MOUSEBUTTONDOWN:
-                if play_button.checkForInput(mouse_pos):
+                if play_button.check_clicked(mouse_pos):
                     pygame.mixer.music.stop()
                     return 1
-                if options_button.checkForInput(mouse_pos):
+                if options_button.check_clicked(mouse_pos):
                     options_menu(SCREEN)
-                if quit_button.checkForInput(mouse_pos):
+                if quit_button.check_clicked(mouse_pos):
                     run = False
         pygame.display.update()
         clock.tick(FPS)
@@ -87,9 +77,9 @@ def options_menu(SCREEN: pygame.Surface):
     bg_image.set_alpha(128)
     
 
-    next_arrow_image = pygame.image.load(os.path.join("assets", "images", "next_arrow.png"))
-    previous_arrow_image = pygame.image.load(os.path.join("assets","images","previous_arrow.png"))
-    return_arrow_image = pygame.transform.scale(pygame.image.load(os.path.join("assets","images","return.png")), (50,50))
+    previous_arrow_image = pygame.transform.smoothscale(pygame.image.load(os.path.join("assets", "images", "previous_arrow.png")),(100,100))
+    next_arrow_image = pygame.transform.smoothscale(pygame.image.load(os.path.join("assets", "images", "next_arrow.png")),(100,100))
+    return_arrow_image = pygame.transform.smoothscale(pygame.image.load(os.path.join("assets","images","return.png")), (50,50))
     text_font = pygame.font.SysFont("consolas", 50)
 
     run = True
@@ -116,21 +106,21 @@ def options_menu(SCREEN: pygame.Surface):
             if event.type == pygame.QUIT:
                 run = False
             if event.type == pygame.MOUSEBUTTONDOWN:
-                if previous_button.checkForInput(mouse_pos):
+                if previous_button.check_clicked(mouse_pos):
                     current_size_index -= 1
                     if current_size_index < 0:
                         current_size_index = len(window_size_list) - 1
                     new_width, new_height = window_size_list[current_size_index]
                     pygame.display.set_mode((new_width, new_height))
                     pygame.display.get_surface().blit(SCREEN, (0, 0))
-                if next_button.checkForInput(mouse_pos):
+                if next_button.check_clicked(mouse_pos):
                     current_size_index += 1
                     if current_size_index >= len(window_size_list):
                         current_size_index = 0
                     new_width, new_height = window_size_list[current_size_index]
                     pygame.display.set_mode((new_width, new_height))
                     pygame.display.get_surface().blit(SCREEN, (0, 0))
-                if return_button.checkForInput(mouse_pos):
+                if return_button.check_clicked(mouse_pos):
                     return
 
         pygame.display.update()
